Made the Todos user specific

+++ backend/src/controllers/todos.ts
@@ -2,9 +2,12 @@ import { RequestHandler } from "express";
 import todoModel from '../models/todo'
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertisDefined";
 export const getTodos : RequestHandler = async (req , res, next)=>{
+    const authenticatedUserId = req.session.userId;
     try{
-        const todos  = await todoModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+        const todos  = await todoModel.find({userId : authenticatedUserId }).exec();
         res.status(200).json(todos);
     }catch(error){
         next(error);
@@ -13,8 +16,10 @@ export const getTodos : RequestHandler = async (req , res, next)=>{
 
 export const getTodo : RequestHandler = async (req, res, next)=>{
     const todoId = req.params.todoId;
+    const authenticatedUserId = req.session.userId;
 
     try{
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(todoId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -22,6 +27,9 @@ export const getTodo : RequestHandler = async (req, res, next)=>{
         const todo  = await todoModel.findById(todoId).exec();
         if(!todo){
             throw createHttpError(404, "Note not found");
+        }
+        if(!todo.userId.equals(authenticatedUserId) ){
+            throw createHttpError(401, "You cannot access this note");
         }
         res.status(200).json(todo);
     }catch(error){
@@ -39,12 +47,16 @@ export const createTodos : RequestHandler<unknown, unknown, createTodoBody, unkn
     const title = req.body.title;
     const text  = req.body.text;
     const isCompleted = req.body.isCompleted;
+    const authenticatedUserId = req.session.userId;
 
     try{
+        assertIsDefined(authenticatedUserId);
+
         if (!title) {
             throw createHttpError(400, "Note must have a title");
         }
         const newTodo = await todoModel.create({
+            userId : authenticatedUserId,
             title : title,
             text : text,
             isCompleted : isCompleted,
@@ -69,8 +81,12 @@ export const updateTodo : RequestHandler<UpdateNoteParams, unknown, UpdateTodoBo
     const newTitle = req.body.title;
     const newText = req.body.text;
     const updateIsCompleted = req.body.isCompleted;
+    const authenticatedUserId = req.session.userId;
+
 
     try{
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(todoId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -82,6 +98,10 @@ export const updateTodo : RequestHandler<UpdateNoteParams, unknown, UpdateTodoBo
         
         if(!todo){
             throw createHttpError(404, "Note not found");
+        }
+
+        if(!todo.userId.equals(authenticatedUserId) ){
+            throw createHttpError(401, "You cannot access this note");
         }
 
         todo.title = newTitle;
@@ -97,8 +117,11 @@ export const updateTodo : RequestHandler<UpdateNoteParams, unknown, UpdateTodoBo
 
 export const deleteTodo : RequestHandler = async (req, res, next)=>{
     const todoId = req.params.todoId;
+    const authenticatedUserId = req.session.userId;
 
     try{
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(todoId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -109,6 +132,10 @@ export const deleteTodo : RequestHandler = async (req, res, next)=>{
             throw createHttpError(404, "Note not found");
         }
 
+        if(!todo.userId.equals(authenticatedUserId) ){
+            throw createHttpError(401, "You cannot access this note");
+        }
+        
         await todo.deleteOne();
         res.sendStatus(204);
     }catch(error){
